config_mgmt: Update user sandbox path
The user's container machine is now named with their uid rather than
username, and the sandbox root directory follows this. We use the
existence of the sandbox root directory to determine whether the
calling user is sandboxed.

Check for the root's parent directory instead, which is unchanged,
as this will work with both the new and old root directory names; and
will not be affected by any future changes.

VRVDR-49808