Add must threshold debug to cfgdbg utility
Part of VRVDR-50300

Add log-type 'must' to the cfgdbg utility, taking an integer value representing time in ms at or above which we wish must statements to be logged so we can monitor performance.

I've reworked the cfgdbg code to allow for String and Int types, keeping the public API the same to avoid code churn.  If you'd prefer a separate / changed API to allow for this instead, let me know.