Fix escaping for backticks
Closes VRVDR-49649

When entering strings on the CLI, anything inside backticks that is not single-quoted is executed by bash, eg 'echo' and 'cat'.  This is expected behaviour, and as the string is executed immediately, this is not a problem.  Where we do have a problem is strings that are single-quoted as the underlying command in back-ticks is stored, and can then be executed later when the config is shown, and if at a higher privilege level, 'bad things' might happen.

This fix ensures that backticks are escaped in the saved config, and for tab completion.  I've added a new testcase to the acceptance tests to verify that backticks inside single quotes work for config save and show, and for exec mode 'show configuration' and 'show configuration commands'.