Create rfc7951 merge helper that understands our schema representation
The generic merge operation in the rfc7951 data library doesn't
understand the schema for the data, as such it does a naive merge on
arrays. When the schema is known one can do better and merge based on
the YANG list key elements. Since we don't want to couple the rfc7951
library to our current schema representation, this is implemented as a
helper in configd for now.

Fixes VRVDR-51009