Tweak memory settings in the configd process
Configd does a lot of operations that use a considerable amount of
temporary memory.  Since we are on a memory constrained system, tweak
the GC settings to ensure it uses less memory than the Go runtime
defaults to.  Instead of waiting for 100% more memory than is
currently allocated before forcing a GC do it at 25%. This impacts
performance slightly but the memory usage tradeoff is worth the
overhead.

Fixes VRVDR-52239