edit_config: default-operation none may merge cfg
Previously an edit-config operation would merge the provided
configuration when "default-operation" was specified as "none", and
no "operation" attribute was specified on any config element.

The expected behaviour in this scenario is for no changes to be made
to the target. From RFC 6241 section 7.2:

> none:  The target datastore is unaffected by the configuration
>    in the <config> parameter, unless and until the incoming
>    configuration data uses the "operation" attribute to request
>    a different operation.

The TestEditConfigDefOpNone test added here triggers this condition,
and verifies the fix in edit_op.Set(). Additional unit tests are also
added to assert the expected behaviour of other default-operation and
operation attribute interactions.

The key change to address the issue is the removal of the call to
edit_op.Merge() when edit_op.op is op_notset and
edit_config.DefaultOperation is def_notset. The additional type
change of edit_config.DefaultOperation is not required but I think
makes the code more obvious; indeed the previous logic in
edit_op.Set() looks to have assumed the type was already
default_operation (through the use of defop_* over op_*).

VRVDR-52424