edit_config.go: Skip accounting on test only
When we are only doing a test operation we should not perform any
accounting, since we are not actually making any changes to the
session.

Although this is a change in behaviour, the previous behaviour was
confusing as it appeared changes were being made to the session when
they were not.

To verify the new behaviour the tests in edit_config_auth_test.go are
extended to verify the expected command authorization and accounting
requests for each of the support operations: testonly, set, testset.

This change is also needed to add hooks for command start accounting.

VRVDR-51580