cli_format.go: Avoid "[] is not valid" in RPC err
When most RPC related errors are formatted by FormatRpcPathError()
they start with "[] is not valid", eg.

    [] is not valid

    RPC failure: TACACS+ daemon is not running

This is noisy and a little confusing since there is no indication of
what might be invalid.

The empty bracket pair is caused by MgmtError.Path being the empty
string. When the MgmtError is encoded as JSON/XML the path is omitted
when empty, so it seems reasonable to do something similar when
formatting such an error for display.

VRVDR-52605