config_mgmt: remove spawnCommandAsCallerFn
The spawnCommandAsCallerFn indirection, and the associated setters in
config_mgmt_internal_test.go, were added to allow the command spawning
function to be changed out for unit tests since the `lu` utility is
typically not present in the UT environment.

To allow UTs to run more production code, the spawnCommandAsCallerFn
indirection is replaced with a flag which can be set to prevent
spawnCommandAsCaller() from executing commands with `lu`.

UTs can now use `SetCallerCmdSetPrivs(false)` to prevent the use of
`lu`. The GetProductionSetPrivs() function is also added to retrieve
the default behaviour, allowing UTs to revert any changes; it is also
used in a basic test to verify the default production behaviour.

VRVDR-53112