session: Add concept of session "owner"
Some sessions are designed to be shared between users, others are
destined to be used only by the creator. Therefore we add the ability
to set an owner for a given session, allowing us to determine if a
given session is shared or owned by a given user.

The API is:

  * OwnedBy(uint32)   - check if the session is owned by a UID
  * IsShared()        - check if a session is shared
  * WithOwner(uint32) - returns an option function for use with
                        NewSession() which sets the session owner
                        to the given UID.

By default, a session is shared (has no owner).

VRVDR-53078

(cherry picked from commit 69fefa65bbf0fded7af4d9249f989f056be251c3)