server: mark "superuser" connections in Context
Update the Context struct with an additional Superuser boolean field
which indicates that the requesting user is either root (UID == 0)
or a member of a given group.

The group can be provided by the "-supergroup" argument to configd,
and the specified value is stored in a new SuperGroup field on the
Config struct.

No group is specified by default, in which case only connections by
root will be marked as a "superuser" connection.

On DANOS, configd is started (via configd.service) with -supergroup
specified as "vyattasu" - the group representing all "superuser"
level users.

This gives an easy method of implementing alternative behaviours
for "superusers" elsewhere in configd.

VRVDR-53078