session: Allow sessions to be shared or un-shared
Update SessionMgr.Create() with an additional boolean argument which
allows the created session to be either shared between users or owned
by the requesting user.

Generally sessions are not shared other than some special cases such
as for NETCONF, and the RUNNING and EFFECTIVE sessions. Un-shared
sessions trigger the access controls implemented in the parent commit.

VRVDR-53078