configd: Add SetupSessionShared RPC
Previously the SetupSession RPC would create a shared session when
deemed necessary (ie. the NETCONF special case).

It should really be up to the client whether a session is shared.
Therefore the special case is removed from SetupSession, and this RPC
now only creates un-shared sessions.

SetupSessionShared should be used when a shared session is required
eg. by a NETCONF client.

VRVDR-53078