session: Add functional options for NewSession()
Add SessionOption, defining a option function signature for the
private session type, and allow a number of options to be passed
to the NewSession() constructor.

The options accept a pointer to the internal session type so that
external users can generate functions to set options, but cannot
execute them without co-operation of a function in the session
library (eg. such as the NewSession() constructor). We can therefore
control in which states options can be applied.

VRVDR-53078

(cherry picked from commit 025f3ab7c62836e3d571e6dbfe84fbd41105d0ce)