Add CopyConfig API
Part of VRVDR-46458.

This commit adds the CopyConfig API that is used for the NETCONF copy-config operation.  Under the covers it is almost identical to Load() except that the input is XML-encoded config instead of the 'configd show' format used by Load().  Thus the code in the session directory sits neatly in load.go.

Most of the unit tests are in the dispatcher (server) directory as the outputChecker methods make it very easy to test.  I needed to add a method for checking a single MgmtError against expectation to the golang-github-danos-mgmterror repo, which is pulled in by the previous commit that ups the required version in the control file.  One test is in session, as it requires more careful handling of permissions, and also very closely matches a load test and hence it put in load_test.go.

Separate commits will add the api to vyatta-cfg, then use it in netconfd, with further unit tests in the latter, and also acceptance tests in QA/auto.  However, the testing here shows that assuming a NETCONF test makes it to the 'dispatcher', it should be handled correctly.

(cherry picked from commit d39b6c299d6ee295d590776651717a422b4e2a36)