Add AAA authorization wrapper to CopyConfig.
Part of VRVDR-46458.  Duncan spotted this was missing - have added copy-config, modelled as 'load encoding xml netconf-copy-config' which maps onto how I plan to implenent load from xml/rfc7951/json.  The existing code for load authorization is extended to allow for encoding as well as routing-instance.

Testing goes as far as (a) nothing fails UT and (b) I'm checking that the 'auther' has the expected command in the log / fails with appropriate error if command is blocked (support in golang-github-danos-config needed for this).  I've not tested with a real TACACS+ server as that's a one-off and the UT covers the pass/fail options anyway.

(cherry picked from commit 01fa2d99bd385519185ca2d8f344bde2f4b4c85d)