Fix crash in expandPath()
Fixes VRVDR-53854.

Softlayer experienced a panic in expandPath.  From code inspection, the problem must have occurred when cfgcli was run with a prefix that didn't match the command, and that was longer than the relevant element of the command's path.  The customer was not actively configuring at the time, so the assumption is that a script must have invoked cfgcli in the background.  It's not clear from the logs, nor from searching BitBucket, what this might have been.  So, the fix addresses the crash, but does not address directly the real root cause.  If the script misbehaves in future, cfgcli will return an error, which will hopefully lead us to eventually detecting the problem application without taking down the customer's router.

NB: I am happy to explicitly log this error in the journal if reviewers think that would be sensible in terms of helping detect future instances of the problem

NB(2): The message format ('Configuration path: path tX [node] has invalid prefix 'to') is going to be somewhat cryptic to a customer, but obvious to us.  Given this seems to be occurring due to a script rather than user action, we don't expect the customer to see this on the command line (where we have enough testing to be reasonably sure the prefix is going to be correct), so I think it's ok, but if you have suggestions for improved wording, let me know.

In terms of testing, the Unit Test crashes with the same traceback as the reported bug without the fix in place.

(cherry picked from commit 89131adeee298ed18f8738014904ca2e0781c0d5)