Move configd flag.Parse() from init() to main()
Seems that somewhere between the Go version in Deb10 and Deb11, running tests now invokes the init() function. This was causing the cmd/configd unit test to fail as it was calling flag.Parse() here and misinterpreting the arguments that were meant for the test harness/infra. flag.Parse() should be called from main() instead.

I think the reason it's in init() is so the logging there can use params passed on the command line, but given main is called directly after, it *ought* to be ok to just parse flags and init logging at the start of main(). In theory if anyone imports this module then init() is run but not main() so this would be a change in behaviour. However, as configd is a standalone executable that is not imported anywhere, that should be a moot point.

I've run a Deb10 ISO with latest configd code (verified with dpkg -s configd) through our systemd and external_components and basic_cli tests and it passes, and checked that the journal entries look ok (given I've relocated when the logger is set up).