Switch DBUS calls to VCI, using ComponentManager.
This commit switches configd over from direct DBUS calls (using code in golang-github-danos-yangd) to using VCI calls. Two new VCI calls (in vci repo) have been added for validate and commit operations. Configd now has its own component, and the VCI client from this is used to create a ComponentManager object that encapsulates the VCI client and a ServiceManager for communicating with systemd. I had tried keeping Dial() and Close() out of the OperationsManager by doing this on configd startup, but it didn't seem to work - somehow the connection was getting broken and the acceptance tests were failing. So for now I have kept Dial and Close in the interface, meaning we need to wrap the client in a local ConfigdOperationsManager object. Can revisit this later. This ComponentManager is initially put in the configd.Context (open to alternative suggestions), and passed into the session further down the calling stack.

The TestSpec in the unit tests (sessiontest) can now create a test ComponentManager which allows us to set and inspect config and state for test components, thus testing the complete calling path. Existing tests validate all 4 APIs (validate, commit, get state and get config) and these continue to pass. component_config_test needed some rework to set/retrieve config from the compmgrtest OpsMgr and SvcMgr so that we really are testing the functionality correctly.

I removed validateFullTree from session_state.go as it's not used, and would have needed modification to keep compiling. component_test.go was so gutted that the remaining few lines were better relocated into submodule_config_test and component_config_test.

Future changes will pull the VCI calls out of the modelset, but this is a good point to pause and get a review. The ComponentManager object may need to move out of the VCI repo into golang-github-danos-config depending on exactly how it ends up interacting with the modelset extension. We can now remove golang-github-danos-yangd from the build.

Part of VRVDR-50443.