Move logic for getting component state to modelset_extension method.
ServiceValidation, ServiceSetRunning and ServiceGetRunning are all (for now) in modelset_extension.go. The 'get state' logic is an outlier, living in session_internal.go, in a different repo (configd). Prior to rejigging configd's use of VCI, it makes sense to relocate this logic into a new ServiceGetState() method in modelset_extension.go. This half of the commit removes the code from configd, including the rfc7951 code associated with it.

Part of VRVDR-50443.

NB: there were no tests that cover this part of the logic. I've written some that verify it works, but without changes coming in a future commit, it's hard to back port here, so they will appear in due course.