Move componentMappings to component manager
Convert to using new componentMappings object, and other general rework in yangd / configd to match changes made to modelset and componentManager objects.

Note that component validation errors are kept as non-fatal as otherwise we see VRVDR-55293 due to multiple usage of dev/null for components with no real config file.

Part of VRVDR-50443.