Fix go vet invocation
Part of VRVDR-56019.

When previously 'fixed' to work with Debian11, the sed statement included a typo which meant that the 'go vet' statement didn't run properly, but also didn't stop the build. Turns out that we don't need the sed statement part anyway, so have just removed it. Instead, we need to only include directories under the _build directory or we get false positives relating to not finding methods exported for test only (eg in export_for_test.go).