Fix over-optimised leafref allowed values cacheing.
Fixes VRVDR-56180.

Earlier optimisation for GPC config that contained lots of rules turns out to have a bug where a leafref with a relative path wholly inside a higher level list is used. We cache values from one list entry, then potentially use them when we are actually wanting to lookup values for a different list entry.

While the GPC is only for the J2 which is not yet shipping, there are other leafrefs that may fall foul of this problem and not allow valid config. Therefore we need to remove optimisation where the leafref path is a relative path. This is detected simply by looking for '..' in the path.

Note that this now means the config that the optimisation was introduced for will be back to original poor performance. That however is only on the non-shipping J2 platform, so is not a problem right now, whereas the risk of invalidating existing configs when shipping Jarrow is high. Note that to reduce time taken for session unit tests I reduced the number of rules in the test for the problem config that was being optimised now that it isn't optimised.

Jarrow RC is being created later today so a speedy review would be appreciated.

(cherry picked from commit 65b787c1c14463b438e1cbf44e6ddeaa2cbbd1c8)